requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
Django==5.0
djangorestframework==3.14.0
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
moviepy==1.0.3
numpy==1.26.2
Pillow==10.1.0
proglog==0.1.10
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tqdm==4.66.1
tzdata==2023.3
urllib3==2.1.0
